+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/topbar/TopBarUrlSplits.kt
@@ -1,9 +1,8 @@
 @file:Suppress("FILE_NAME_MATCH_CLASS")
 
-package com.saveourtool.save.frontend.common.components.topbar
+package com.saveourtool.save.cosv.frontend.components.topbar
 
 import com.saveourtool.save.frontend.common.utils.TopBarUrl
-import com.saveourtool.save.frontend.common.utils.isCosvDomain
 import com.saveourtool.save.utils.URL_PATH_DELIMITER
 import com.saveourtool.save.validation.FrontendRoutes
 
@@ -19,8 +18,6 @@ import react.dom.html.ReactHTML.ol
 import react.router.dom.Link
 import remix.run.router.Location
 import web.cssom.ClassName
-
-import kotlinx.browser.window
 
 /**
  * Displays the URL split with "/".
@@ -39,7 +36,7 @@ val topBarUrlSplits: FC<TopBarUrlSplitsProps> = FC { props ->
                     // if we are on welcome page right now - need to highlight SAVE in menu
                     val textColor = if (props.location.pathname == "/") "text-warning" else "text-light"
                     className = ClassName(textColor)
-                    +if (window.location.isCosvDomain()) "COSV" else "SaveOurTool!"
+                    +"COSV"
                 }
             }
             props.location.pathname
