+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/VulnerabilityGeneralInfoProps.kt
@@ -338,7 +338,7 @@ val vulnerabilityGeneralInfoProps: FC<VulnerabilityGeneralInfoProps> = FC { prop
                         +"Contributors".t()
                     }
                     userBoard {
-                        users = saveContributors
+                        users = saveContributors.distinctBy { it.name }
                         avatarOuterClasses = "col-2"
                     }
                 }
@@ -351,8 +351,8 @@ val vulnerabilityGeneralInfoProps: FC<VulnerabilityGeneralInfoProps> = FC { prop
                         className = ClassName("font-weight-bold text-primary-blue mb-4")
                         +"Credits"
                     }
-                    cosv.credits!!.forEachIndexed { creditInd, credit ->
-                        val userPrefix = if (credit.name in saveContributors.map { it.name }) "[SAVE] user" else "External user"
+                    cosv.credits!!.mergeContactsForOneUser().forEachIndexed { creditInd, credit ->
+                        val userPrefix = if (credit.name in saveContributors.map { it.name }) "[SAVE user]" else "[External user]"
                         if (props.isEditDisabled) {
                             div {
                                 className = ClassName("row ml-1")
@@ -517,14 +517,22 @@ external interface VulnerabilityGeneralInfoProps : Props {
     var setSelectedMenu: StateSetter<VulnerabilityTab>
 }
 
+private fun List<Credit>.mergeContactsForOneUser() = this
+    .groupBy { it.name }.map { entry ->
+        val name = entry.key
+        val contacts = entry.value.mapNotNull { it.contact }.flatten()
+        // suppose, that `credit type` is the same for credits with same name
+        Credit(name, contacts, entry.value.first().type)
+    }
+
 /**
  * @param currentUserInfo
  * @param vulnerability
  * @return if user can press edit button
  */
 private fun hasRightsToEdit(currentUserInfo: UserInfo?, vulnerability: VulnerabilityExt): Boolean =
-        currentUserInfo?.isSuperAdmin() == true || ((currentUserInfo?.name == vulnerability.metadataDto.user.name ||
-                currentUserInfo?.name in vulnerability.getAllParticipants().map { it.name }) && vulnerability.metadataDto.status != VulnerabilityStatus.APPROVED)
+    currentUserInfo?.isSuperAdmin() == true || ((currentUserInfo?.name == vulnerability.metadataDto.user.name ||
+            currentUserInfo?.name in vulnerability.getAllParticipants().map { it.name }) && vulnerability.metadataDto.status != VulnerabilityStatus.APPROVED)
 
 private fun updateCredits(
     cosvCredits: List<Credit>,
@@ -537,3 +545,4 @@ private fun updateCredits(
     updatedCredits[currentCreditInd] = updatedCredit
     return updatedCredits
 }
+
