+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/VulnerabilityInfoTab.kt
@@ -30,6 +30,7 @@ import com.saveourtool.save.utils.getTimeline
 import com.saveourtool.save.utils.listToShortString
 
 import com.saveourtool.osv4k.Affected
+import com.saveourtool.save.utils.ELLIPSIS
 import react.*
 import react.dom.html.ReactHTML.div
 import react.dom.html.ReactHTML.h4
@@ -301,6 +302,7 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
     @Suppress("COMPLEX_EXPRESSION")
     renderProjects(
         "Affected projects".t(),
+        "affectedCarousel",
         VulnerabilityProjectType.PROJECT,
         vulnerabilityProjects,
         projectTable,
@@ -314,6 +316,7 @@ val vulnerabilityInfoTab: FC<VulnerabilityInfoTabProps> = FC { props ->
     @Suppress("COMPLEX_EXPRESSION")
     renderProjects(
         "Commits with fix".t(),
+        "affectedCarousel",
         VulnerabilityProjectType.COMMIT,
         vulnerabilityProjects,
         commitFixTable,
@@ -391,6 +394,7 @@ external interface VulnerabilityInfoTabProps : Props {
 )
 private fun ChildrenBuilder.renderProjects(
     sectionName: String,
+    carouselBodyId: String,
     projectType: VulnerabilityProjectType,
     projects: List<CosvAffected>,
     table: FC<AffectedProjectTableProps>,
@@ -418,7 +422,7 @@ private fun ChildrenBuilder.renderProjects(
             } else {
                 carousel(
                     projects.toList(),
-                    "carousel",
+                    carouselBodyId,
                     outerClasses = "border border-secondary",
                 ) {
                     renderProjectCard(it, projectType, t)
@@ -470,7 +474,7 @@ private fun ChildrenBuilder.renderVersions(versions: List<String?>) {
             forEach { span { +it } }
         } else {
             span { +first() }
-            +" ... "
+            +" $ELLIPSIS "
             span { +last() }
         }
     }
