+++ save-cosv-frontend/src/main/kotlin/com/saveourtool/save/cosv/frontend/components/views/vuln/VulnerabilityTagsComponent.kt
@@ -48,7 +48,7 @@ val vulnerabilityTagsComponent: FC<VulnerabilityTagsComponentProps> = FC { props
                 className = ClassName("form-control custom-input $validityClassName")
                 value = newTag
                 placeholder = "Add a new tag...".t()
-                title = "Tag should not have commas, length should be more than 2 and less than 16.".t()
+                title = "Tag should not have commas, length should be more than 1 and less than 16.".t()
                 asDynamic()["data-toggle"] = "tooltip"
                 asDynamic()["data-placement"] = "top"
                 onChange = { event -> setNewTag(event.target.value) }
